app/requirements.txt
google-cloud-bigquery
google-cloud-storage
google-cloud-vision
pandas==1.4.2
Flask
tensorflow==2.8.0
Pillow==9.1.0
